+++ src/components/preconsult/preconsult-page.js
@@ -6,6 +6,7 @@ import FillOutForm from "./fill-out-form"
 import Queued from "./queued"
 import InConsult from "./in-consult"
 import ProceedWithProbing from "./proceed-with-probing"
+import Reassess from "./reassess"
 
 const PreconsultPage = ({ paths }) => {
   const { state, dispatch } = useContext(AppContext)
@@ -14,7 +15,7 @@ const PreconsultPage = ({ paths }) => {
     switch (state?.preConsult?.status) {
       case "VISIT_CLINIC":
       default:
-        return <VisitClinic state={state} dispatch={dispatch} />
+        return <VisitClinic dispatch={dispatch} />
       case "FILL_OUT_FORM":
         return <FillOutForm paths={paths} state={state} dispatch={dispatch} />
       case "QUEUED":
@@ -24,6 +25,8 @@ const PreconsultPage = ({ paths }) => {
       case "PROCEED_WITH_PROBING":
       case "FINISH_PRECONSULT":
         return <ProceedWithProbing state={state} dispatch={dispatch} />
+      case "REASSESS":
+        return <Reassess paths={paths} state={state} dispatch={dispatch} />
     }
   }