+++ src/pages/index.js
@@ -3,7 +3,7 @@ import React from "react"
 
 import Layout from "../components/layout/layout"
 import Seo from "../components/layout/seo"
-import PreconsultPage from "../components/preconsult/preconsult"
+import PreconsultPage from "../components/preconsult/preconsult-page"
 
 const IndexPage = props => {
   return (