+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Prescribe Me Books Changelog
 
+# v0.0.2
+
+- Added probing paths and next/back navigation
+- Added prescription page
+- Add a wildcard reassessment page
+
 # Initial Release (v0.0.1)
 
 - Added AppContext